+++ changelog.md
@@ -9,9 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Add test mode - `rake db_validator:test`
+- Add error count and grouping in JSON reports
+- Save JSON report to a file in the `db_validator_reports` directory
 
 ### Fixed
 - Skip validation of HABTM (Has And Belongs To Many) join tables
+
+### Changed
+- Restructure JSON output format to include error counts and grouped records
 
 ## [0.2.0] - 2024-11-07